Update app.css
adding mobile support, so you can see the patches in a right format on your mobile as well.

Only thing could be creating a new patch from a mobile device. This can be fixed on a later state. 
As having the ability to see the patches on your mobile is a really nice way already. :)